Fixed hover bug
There was a bug where if you held both space and ctrl down, after the initial jump you would "hover" in an odd way(but fun). Found out this had to do with the game thinking the player was climbing a wall every time he crouched, so I came up with a fix for it!